+++ src/app/containers/MostRead/Canonical/index.jsx
@@ -1,31 +1,12 @@
 import React, { useEffect, useState, useContext } from 'react';
 import 'isomorphic-fetch';
 import { bool, string } from 'prop-types';
-import styled from 'styled-components';
-import {
-  GEL_GROUP_1_SCREEN_WIDTH_MAX,
-  GEL_GROUP_2_SCREEN_WIDTH_MIN,
-  GEL_GROUP_3_SCREEN_WIDTH_MIN,
-  GEL_GROUP_3_SCREEN_WIDTH_MAX,
-  GEL_GROUP_4_SCREEN_WIDTH_MIN,
-  GEL_GROUP_4_SCREEN_WIDTH_MAX,
-  GEL_GROUP_5_SCREEN_WIDTH_MIN,
-} from '@bbc/gel-foundations/breakpoints';
-import {
-  GEL_MARGIN_ABOVE_400PX,
-  GEL_MARGIN_BELOW_400PX,
-  GEL_SPACING_DBL,
-  GEL_SPACING_TRPL,
-  GEL_SPACING_QUAD,
-  GEL_SPACING_QUIN,
-} from '@bbc/gel-foundations/spacings';
 import {
   MostReadList,
   MostReadItemWrapper,
   MostReadRank,
   MostReadLink,
 } from '@bbc/psammead-most-read';
-import SectionLabel from '@bbc/psammead-section-label';
 import { ServiceContext } from '#contexts/ServiceContext';
 import webLogger from '#lib/logger.web';
 import { mostReadRecordIsFresh, shouldRenderLastUpdated } from '../utilities';
@@ -33,54 +14,7 @@ import LastUpdated from './LastUpdated';
 
 const logger = webLogger();
 
-const MarginWrapper = styled.div`
-  @media (min-width: ${GEL_GROUP_3_SCREEN_WIDTH_MIN}) {
-    margin-top: ${GEL_SPACING_DBL};
-  }
-  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) {
-    margin-top: ${GEL_SPACING_TRPL};
-  }
-`;
-
-const MostReadSection = styled.section.attrs(() => ({
-  role: 'region',
-  'aria-labelledby': 'Most-Read',
-  'data-e2e': 'most-read',
-}))``;
-
-const FrontPageMostReadSection = styled(MostReadSection)`
-  /* To centre page layout for Group 4+ */
-  margin: 0 auto;
-  width: 100%; /* Needed for IE11 */
-  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) {
-    max-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN};
-  }
-`;
-
-const ConstrainedMostReadSection = styled(MostReadSection)`
-  @media (max-width: ${GEL_GROUP_1_SCREEN_WIDTH_MAX}) {
-    margin: 0 ${GEL_MARGIN_BELOW_400PX} ${GEL_SPACING_TRPL};
-  }
-  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) and (max-width: ${GEL_GROUP_3_SCREEN_WIDTH_MAX}) {
-    margin: 0 ${GEL_MARGIN_ABOVE_400PX} ${GEL_SPACING_QUAD};
-  }
-  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) and (max-width: ${GEL_GROUP_4_SCREEN_WIDTH_MAX}) {
-    margin: 0 ${GEL_MARGIN_ABOVE_400PX} ${GEL_SPACING_QUIN};
-  }
-  @media (min-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN}) {
-    width: 100%; /* Needed for IE11 */
-    margin: 0 auto ${GEL_SPACING_TRPL};
-    padding: 0 ${GEL_SPACING_DBL};
-    max-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN};
-  }
-`;
-
-const CanonicalMostRead = ({
-  endpoint,
-  maxTwoColumns,
-  constrainMaxWidth,
-  isOnFrontPage,
-}) => {
+const CanonicalMostRead = ({ endpoint, maxTwoColumns }) => {
   const [items, setItems] = useState([]);
   const {
     service,
@@ -88,7 +22,7 @@ const CanonicalMostRead = ({
     dir,
     datetimeLocale,
     timezone,
-    mostRead: { header, lastUpdated, numberOfItems },
+    mostRead: { lastUpdated, numberOfItems },
   } = useContext(ServiceContext);
 
   useEffect(() => {
@@ -136,75 +70,54 @@ const CanonicalMostRead = ({
     timezone,
   ]);
 
+  console.log(items);
+
   if (!items.length) {
     return null;
   }
 
-  const StyledMostRead = isOnFrontPage
-    ? FrontPageMostReadSection
-    : MostReadSection;
-
-  const MostReadSectionWrapper = constrainMaxWidth
-    ? ConstrainedMostReadSection
-    : StyledMostRead;
-
   return (
-    <MostReadSectionWrapper>
-      <SectionLabel
-        script={script}
-        labelId="Most-Read"
-        service={service}
-        dir={dir}
-      >
-        {header}
-      </SectionLabel>
-      <MarginWrapper>
-        <MostReadList
-          numberOfItems={items.length}
+    <MostReadList
+      numberOfItems={items.length}
+      dir={dir}
+      maxTwoColumns={maxTwoColumns}
+    >
+      {items.map((item, i) => (
+        <MostReadItemWrapper
           dir={dir}
+          key={item.id}
           maxTwoColumns={maxTwoColumns}
         >
-          {items.map((item, i) => (
-            <MostReadItemWrapper
-              dir={dir}
-              key={item.id}
-              maxTwoColumns={maxTwoColumns}
-            >
-              <MostReadRank
-                service={service}
-                script={script}
-                listIndex={i + 1}
-                numberOfItems={items.length}
-                dir={dir}
-                maxTwoColumns={maxTwoColumns}
-              />
-              <MostReadLink
-                dir={dir}
-                service={service}
-                script={script}
-                title={item.title}
-                href={item.href}
-              >
-                {item.timestamp}
-              </MostReadLink>
-            </MostReadItemWrapper>
-          ))}
-        </MostReadList>
-      </MarginWrapper>
-    </MostReadSectionWrapper>
+          <MostReadRank
+            service={service}
+            script={script}
+            listIndex={i + 1}
+            numberOfItems={items.length}
+            dir={dir}
+            maxTwoColumns={maxTwoColumns}
+          />
+          <MostReadLink
+            dir={dir}
+            service={service}
+            script={script}
+            title={item.title}
+            href={item.href}
+          >
+            {item.timestamp}
+          </MostReadLink>
+        </MostReadItemWrapper>
+      ))}
+    </MostReadList>
   );
 };
 
 CanonicalMostRead.propTypes = {
   endpoint: string.isRequired,
-  constrainMaxWidth: bool.isRequired,
   maxTwoColumns: bool,
-  isOnFrontPage: bool,
 };
 
 CanonicalMostRead.defaultProps = {
   maxTwoColumns: false,
-  isOnFrontPage: false,
 };
 
 export default CanonicalMostRead;
